Refactored code in controller to service layer.
